Mm 34441 Cloud Trial in subscription screen (#7834)
* MM-34441 - updates in subscription screen

* add changes based on design updates

* more updates based on ux changes

* fix to contact sales card

* fix checks issues

* fix styles and i18n checks

* add code changes based on PR comments

* fix trialing word and TEMP use of end_at instead of trial_end_at

* improvements based on Pr comments

* use the right value for the trial-end-at
