+++ i18n/en.json
@@ -268,6 +268,12 @@
   "admin.billing.subscription.creditCardHasExpired.description.updatePaymentInformation": "update your payment information",
   "admin.billing.subscription.creditCardHasExpired.please": "Please ",
   "admin.billing.subscription.disclaimer": "Your total is calculated at the end of the billing cycle based on the number of enabled users. You’ll only be charged if you exceed the free tier limits.",
+  "admin.billing.subscription.freeTrial.description": "Your free trial will expire in {daysLeftOnTrial} days. Add your payment information to continue after the trial ends.",
+  "admin.billing.subscription.freeTrial.lastDay.description": "Your free trial has ended. Add payment information to continue enjoying the benefits of Cloud Professional.",
+  "admin.billing.subscription.freeTrial.lastDay.title": "Your free trial ends today",
+  "admin.billing.subscription.freeTrial.lessThan3Days.description": "Your free trial will end in {daysLeftOnTrial, number} {daysLeftOnTrial, plural, one {day} other {days}}. Add payment information to continue enjoying the benefits of Cloud Professional.",
+  "admin.billing.subscription.freeTrial.subscribeButton": "Subscribe Now",
+  "admin.billing.subscription.freeTrial.title": "You're currently on a free trial",
   "admin.billing.subscription.goBackTryAgain": "Go back and try again",
   "admin.billing.subscription.howItWorks": "See how billing works",
   "admin.billing.subscription.letsGo": "Lets go!",
@@ -283,17 +289,22 @@
   "admin.billing.subscription.planDetails.endDate": "End Date: ",
   "admin.billing.subscription.planDetails.features.10GBstoragePerUser": "10 GB storage per user",
   "admin.billing.subscription.planDetails.features.99uptime": "99.0% uptime",
+  "admin.billing.subscription.planDetails.features.advanceTeamPermission": "Advanced team permissions",
+  "admin.billing.subscription.planDetails.features.channelModeration": "Channel moderation",
   "admin.billing.subscription.planDetails.features.guestAccounts": "Guest Accounts",
   "admin.billing.subscription.planDetails.features.mfaAuthentication": "Google, Gitlab, O365 & MFA Authentication",
+  "admin.billing.subscription.planDetails.features.mfaEnforcement": "MFA enforcement",
+  "admin.billing.subscription.planDetails.features.multiplatformSso": "Gitlab, Google, and O365 single sign-on",
+  "admin.billing.subscription.planDetails.features.readOnlyChannels": "Read-only announcement channels",
   "admin.billing.subscription.planDetails.features.selfServiceDocumentation": "Self-Service documentation and forum support",
   "admin.billing.subscription.planDetails.features.unlimitedIntegrations": "Unlimited Integrations",
   "admin.billing.subscription.planDetails.features.unlimitedTeamsAndChannels": "Unlimited teams, channels, and search history",
   "admin.billing.subscription.planDetails.howBillingWorks": "See how billing works",
   "admin.billing.subscription.planDetails.numberOfSeats": "{numberOfSeats} seats",
   "admin.billing.subscription.planDetails.numberOfSeatsRegistered": "({userCount} currently registered)",
-  "admin.billing.subscription.planDetails.perUserPerMonth": "/user/month",
+  "admin.billing.subscription.planDetails.perUserPerMonth": "/user/month.",
   "admin.billing.subscription.planDetails.planDetailsName.freeForXOrMoreUsers": "/user/month for {aboveUserLimit} or more users.",
-  "admin.billing.subscription.planDetails.planDetailsName.freeUpTo": "Free for up to {aboveUserLimit} users.",
+  "admin.billing.subscription.planDetails.productName.cloudProfessional": "Cloud Professional",
   "admin.billing.subscription.planDetails.prolongedOverages": "Prolonged overages may result in additional charges.",
   "admin.billing.subscription.planDetails.seatCountOverages": "Seat count overages",
   "admin.billing.subscription.planDetails.startDate": "Start Date: ",
@@ -301,10 +312,12 @@
   "admin.billing.subscription.planDetails.upToXUsers": "up to {userLimit} users",
   "admin.billing.subscription.planDetails.userCount": "{userCount} users",
   "admin.billing.subscription.planDetails.userCountWithLimit": "{userCount} / {userLimit} users",
+  "admin.billing.subscription.privateCloudCard.cloudProfessional.description": "Optimize your processes with VPC Peering, a dedicated AWS account and premium support.",
+  "admin.billing.subscription.privateCloudCard.cloudProfessional.title": "Upgrade to Cloud Enterprise",
   "admin.billing.subscription.privateCloudCard.contactSales": "Contact Sales",
   "admin.billing.subscription.privateCloudCard.contactSupport": "Contact Support",
-  "admin.billing.subscription.privateCloudCard.description": "If you need software with dedicated, single-tenant architecture, Mattermost Private Cloud (Beta) is the solution for high-trust collaboration.",
-  "admin.billing.subscription.privateCloudCard.title": "Looking for a high-trust private cloud?",
+  "admin.billing.subscription.privateCloudCard.freeTrial.description": "We love to work with our customers and their needs. Contact sales for subscription, billing or trial-specific questions.",
+  "admin.billing.subscription.privateCloudCard.freeTrial.title": "Questions about your trial?",
   "admin.billing.subscription.questions": "Questions?",
   "admin.billing.subscription.stateprovince": "State/Province",
   "admin.billing.subscription.title": "Subscriptions",
@@ -421,6 +434,7 @@
   "admin.channel_settings.description": "Manage channel settings.",
   "admin.channel_settings.groupsPageTitle": "{siteName} Channels",
   "admin.channel_settings.title": "Channels",
+  "admin.cloud.import.header.TrialBadge": "Trial",
   "admin.cluster.ClusterName": "Cluster Name:",
   "admin.cluster.ClusterNameDesc": "The cluster to join by name.  Only nodes with the same cluster name will join together.  This is to support Blue-Green deployments or staging pointing to the same database.",
   "admin.cluster.ClusterNameEx": "E.g.: \"Production\" or \"Staging\"",
