+++ packages/mattermost-redux/src/types/cloud.ts
@@ -22,6 +22,7 @@ export type Subscription = {
     seats: number;
     is_paid_tier: string;
     last_invoice?: Invoice;
+    status: string;
     trial_end_at: number;
 }
